### 2019-11-23 #15 @sjvrijn
MF2: Multi-Fidelity Functions
Submitting Author:  Sander van Rijn(@sjvrijn)
Package Name:  mf2
One-Line Description of Package:  A collection of multi-fidelity analytical benchmark functions
Repository Link (if existing):  https://github.com/sjvrijn/mf2

---

This package provides a number of multi-fidelity function implementations commonly used to benchmark the performance of multi-fidelity (surrogate-model-assisted) optimization algorithms. It provides a common and simple interface to all included functions, with corresponding references available in the documentation.

- Please indicate which [category or categories](https://pyopensci.github.io/dev_guide/peer_review/peer_review_proc.html#aims-and-scope) this package falls under:

Analytical benchmark functions are a like well-established datasets in the field of machine-learning: by comparing your algorithm on the same ones, you can accurately compare your results. They are essential for _reproducibility_, as typos in the formulas will obviously give very different results, but are not an 'external' utility like the examples given under [Package Categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html#Package-Categories) in the guidebook.

The target audience is any scientist working in the field of multi-fidelity optimization algorithms, as they will likely want to use functions from this collection to benchmark the performance of their algorithms.

I was not aware of any when writing the package. Since recently, there seem to be two packages [`emukit`](https://github.com/amzn/emukit/tree/master/emukit/test_functions) and [`smt`](https://github.com/SMTorg/smt/tree/master/smt/problems) that also contain a partially overlapping set of the same functions. Those packages however are mostly focussed on providing algorithms to apply on these kinds of problems, while my `mf2` is a separate package consisting of just functions.

In my opinion, that makes mf2:
 1. **easier to find** when searching for function implementations,
 2. **easier to include** when working on new or different methods, and
 3. **better tested**  as mf2 has dedicated and comprehensive tests to verify the functions (property and regression tests)

None
