+++ README.md
@@ -2,7 +2,7 @@
 [![Build Status](https://travis-ci.org/sjvrijn/multi-fidelity-functions.svg?branch=master)](https://travis-ci.org/sjvrijn/multi-fidelity-functions)
 [![Coverage Status](https://coveralls.io/repos/github/sjvrijn/multi-fidelity-functions/badge.svg?branch=master)](https://coveralls.io/github/sjvrijn/multi-fidelity-functions?branch=master)
 [![Codacy Badge](https://api.codacy.com/project/badge/Grade/54144e7d406b4558a14996b06a89adf8)](https://www.codacy.com/manual/sjvrijn/multi-fidelity-functions?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=sjvrijn/multi-fidelity-functions&amp;utm_campaign=Badge_Grade)
-[![Documentation Status](https://readthedocs.org/projects/multi-fidelity-functions/badge/?version=latest)](https://multi-fidelity-functions.readthedocs.io/en/latest/?badge=latest)
+[![Documentation Status](https://readthedocs.org/projects/mf2/badge/?version=latest)](https://mf2.readthedocs.io/en/latest/?badge=latest)
 
 # Multi-Fidelity-Functions
 
