### 2019-12-03 183423d
updated borehole documentation
