### 2019-12-04 #15 @lwasser
thank you for this submission @sjvrijn our next pyopensci checkin in is tomorrow at 11am mountain time! we will discuss this package then. you are welcome to join us as well as it's an open meeting!!!
