### 2019-12-09 #15 @sjvrijn
As discussed in the December 5 meeting, this submission is withdrawn/on hold

See [the meeting notes](https://hackmd.io/cHsYlTvwT7eWz8y21v8G5w) for details.
