### 2019-12-10 #15 @lwasser
awesome! and funny you should ask!! i finally created a good form for this very purpose. You can fill it out here -- https://forms.gle/5MuYMVs8xpCxk9w29 and please let me know if you have any issues with filling it out or if there are other "categories" of domain expertise that i am missing!! i am sure i've missed some important ones.
